--- astro/media-manager/wix.config copy.json
@@ -1,4 +1,0 @@
-{
-  "appId": "ab1546dc-6baa-46d4-b6de-fcb1fa5c3a97",
-  "projectId": "9506d982-4cf6-4700-a46e-49465a622ecb"
-}
